Abstract out common code in Inventory and Cart to IngredientList
