+++ src/main/java/seedu/address/model/ReadOnlyCart.java
@@ -1,15 +1,6 @@
 package seedu.address.model;
 
-import javafx.collections.ObservableList;
-import seedu.address.model.ingredient.Ingredient;
-
 /**
  * Unmodifiable view of a shopping cart
  */
-public interface ReadOnlyCart {
-    /**
-     * Returns an unmodifiable view of the persons list.
-     * This list will not contain any duplicate persons.
-     */
-    ObservableList<Ingredient> getIngredientList();
-}
+public interface ReadOnlyCart extends ReadOnlyIngredientList {}
