+++ src/main/java/seedu/address/model/ReadOnlyIngredientList.java
@@ -1,0 +1,15 @@
+package seedu.address.model;
+
+import javafx.collections.ObservableList;
+import seedu.address.model.ingredient.Ingredient;
+
+/**
+ * Unmodifiable view of an ingredient list
+ */
+public interface ReadOnlyIngredientList {
+    /**
+     * Returns an unmodifiable view of the ingredients list.
+     * This list will not contain any duplicate ingredients.
+     */
+    ObservableList<Ingredient> getIngredientList();
+}
