+++ src/main/java/seedu/address/model/ReadOnlyInventory.java
@@ -1,15 +1,6 @@
 package seedu.address.model;
 
-import javafx.collections.ObservableList;
-import seedu.address.model.ingredient.Ingredient;
-
 /**
  * Unmodifiable view of an inventory
  */
-public interface ReadOnlyInventory {
-    /**
-     * Returns an unmodifiable view of the persons list.
-     * This list will not contain any duplicate persons.
-     */
-    ObservableList<Ingredient> getIngredientList();
-}
+public interface ReadOnlyInventory extends ReadOnlyIngredientList {}
